pyproject.toml
[tool.poetry]
name = "flask-restapi"
version = "0.1.5"
description = ""
authors = ["jonars <jonarsli13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = {extras = ["async"], version = "^2.0.1"}
pydantic = "^1.8.2"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.8.0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"